requirements.txt
# File: requirements.txt

fastapi
uvicorn
numpy

# OpenTelemetry for Honeycomb observability
opentelemetry-instrumentation
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-processor-baggage
